// src/app/core/user/user.types.ts
export interface User {
    name: string;
    email: string;
    avatar?: string;
    status?: string;
    realmAccess: any[];
    resourceAccess: any;
    officeId: string;
    orgId: string;
    employeeId: string;
    username: string;
    clickStreamTrack: string;
    userTerminalIP: any[];
}